Add AWS SES emailer project to solution
This commit includes the addition of a new project 'aws-ses-emailer'. This project is dedicated to handling email operations using Amazon's Simple Email Service (SES). Till now, the system depended on SMTP for sending emails which was not reliable enough. SES offers greater deliverability rates and simpler setup hence the change. The new service makes it possible to send an email with attachments.

Multiple files have been added as a part of this project such as 'AWSEmailService.cs', 'EmailAttachment.cs', 'Program.cs', and 'IEmailService.cs' to handle different aspects of the email operation. The .sln file was updated to reflect this new project. Additionally, multiple dependencies needed for the operation of AWS SES have been added to the 'aws-ses-emailer.csproj' file.