Add Equity Award Pool system with actors
This commit introduces a new Equity Award Pool system in the akka-net-sample project. This system is represented with Akka.Net actors for handling the options award workflow. It includes the creation of the EquityAwardPoolActor, AwardOption, HumanResourcesActor, and EmployeeActor classes. The EquityAwardPoolActor is for managing the pool of options, the HumanResourcesActor creates child actor instances for each employee and manages award requests, while the EmployeeActor simply forwards any award requests to the EquityAwardPoolActor. The AwardOption class represents a request to award options to an employee. The akka-net-sample project has been added to csharp-playpen.sln as well. These changes allow the simulation of