Add UTC DateTime converter function
Introduced a new function `GetDateTimeInUtc` to handle conversion of date strings to UTC formatted dates. This change is necessary to normalize date data received in different formats and time zones. The function also validates the input date string and throws an exception with detailed error if the input format is not `yyyy-MM-dd`. Also, commenting out a line invoking `Sha512()` method on a string, as it's currently not in use.