Update target framework to net8.0 in all projects
The TargetFramework in all projects was updated from net7.0 to net8.0 to get advantage of the latest features and performance improvements. This was done by removing specific target framework in each project and specifying it globally in Directory.Build.props. This change also simplifies future updates since we only have to modify one file.