Added OpenTelemetry logging and tracing
This commit introduces OpenTelemetry (OTel) in the application. This change includes adding dependencies related to OTel in the .csproj file and setup of OTel in Program.cs. A dedicated extension method is also created to modularize the OTel setup. This helps in improved observability with logging and tracing features provided by OpenTelemetry.