Add C# optional monad example
Added a new file 'Program.cs' under 'src/csharp-optional-monad/' to demonstrate optional monad usage in C#.
This code illustrates the definition and application of the optional monad pattern in the context of handling null values for author in the 'Book' class and last name in the 'Person' class. The corresponding project file 'csharp-optional-monad.csproj' has also been added and included in the solution file 'csharp-playpen.sln'. This change is made for educational purposes and enhancing code readability.