Add MongoDB LINQ provider project
A new project 'mongo-v3-linq-provider' was added to the solution to facilitate querying MongoDB using LINQ. This includes the Project setup with output type 'Exe' and addition of required packages like 'Bogus' and 'MongoDB.Driver'. The 'Program.cs' sets up a dummy Person model for MongoDB and the 'Person.cs' defines this schema. This is an initial step towards MongoDB integration.